Issue #2807: extracted FxaProfile to its own file
