+++ app/src/test/java/org/mozilla/tv/firefox/fxa/FxaRepoTest.kt
@@ -9,25 +9,21 @@ import io.mockk.MockKAnnotations
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
 import io.mockk.mockk
-import io.mockk.mockkObject
 import io.mockk.verify
 import io.reactivex.Observable
 import io.reactivex.observers.TestObserver
 import io.reactivex.schedulers.TestScheduler
 import io.reactivex.subjects.PublishSubject
 import mozilla.components.concept.sync.AuthType
-import mozilla.components.concept.sync.Avatar
 import mozilla.components.concept.sync.DeviceType
 import mozilla.components.concept.sync.OAuthAccount
 import mozilla.components.concept.sync.Profile
 import mozilla.components.concept.sync.TabData
 import mozilla.components.service.fxa.manager.FxaAccountManager
-import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.BeforeClass
 import org.junit.Test
 import org.mozilla.tv.firefox.R
-import org.mozilla.tv.firefox.channels.ImageSetStrategy
 import org.mozilla.tv.firefox.framework.UnresolvedString
 import org.mozilla.tv.firefox.helpers.RxTestHelper
 import org.mozilla.tv.firefox.telemetry.SentryIntegration
@@ -57,8 +53,6 @@ class FxaRepoTest {
     private lateinit var accountState: Observable<FxaRepo.AccountState>
     private lateinit var accountStateTestObs: TestObserver<FxaRepo.AccountState>
     private lateinit var receivedTabsTestObs: TestObserver<FxaReceivedTab>
-
-    private val defaultProfileAvatarImage = ImageSetStrategy.ById(R.drawable.ic_default_avatar)
 
     @Before
     fun setUp() {
@@ -133,81 +127,6 @@ class FxaRepoTest {
         val profile = Profile("uid", "email", null, "displayName")
         fxaRepo.accountObserver.onProfileUpdated(profile)
         accountStateTestObs.assertValueAt(1) { it is FxaRepo.AccountState.AuthenticatedWithProfile }
-    }
-
-    @Test
-    fun `GIVEN profile has valid displayName WHEN profile is converted to domain object THEN displayName should be displayName`() {
-        val displayName = "displayName"
-
-        val profiles = listOf(
-            Profile(null, null, null, displayName),
-            Profile("uid", null, null, displayName),
-            Profile(null, "email", null, displayName),
-            Profile(null, null, Avatar("url", true), displayName),
-            Profile("uid", "email", Avatar("url", true), displayName)
-        )
-
-        profiles.map { it.toDomainObject() }
-            .map { it.displayName }
-            .forEach { assertEquals(displayName, it) }
-    }
-
-    @Test
-    fun `GIVEN profile has no displayName WHEN profile is converted to domain object THEN displayName should be email`() {
-        val email = "email"
-        val profile = Profile("uid", email, null, null)
-        assertEquals(email, profile.toDomainObject().displayName)
-    }
-
-    @Test
-    fun `GIVEN profile has no displayName or email WHEN profile is converted to domain object THEN displayName should be blank`() {
-        val profile = Profile("uid", null, null, null)
-        assertEquals("", profile.toDomainObject().displayName)
-    }
-
-    @Test
-    fun `GIVEN profile has no displayName or email WHEN profile is converted to domain object THEN sentry should log an error`() {
-        mockkObject(SentryIntegration)
-
-        Profile("uid", null, null, null).toDomainObject()
-        verify(exactly = 1) { SentryIntegration.captureAndLogError(any(), any()) }
-    }
-
-    @Test
-    fun `GIVEN profile has no avatar WHEN profile is converted to domain object THEN default resource image should be used`() {
-        val profile = Profile("uid", null, null, null)
-
-        assertEquals(defaultProfileAvatarImage, profile.toDomainObject().avatarSetStrategy)
-    }
-
-    @Test
-    fun `GIVEN profile avatar is fxa default WHEN profile is converted to domain object THEN default resource image should be used`() {
-        val profiles = listOf(
-            Profile("uid", null, Avatar("https://firefoxusercontent.com/00000000000000000000000000000000", true), null),
-            Profile("uid", null, Avatar("https://firefoxusercontent.com/00000000000000000000000000000000", false), null),
-            Profile(null, null, Avatar("https://firefoxusercontent.com/00000000000000000000000000000000", false), null),
-            Profile("uid", null, Avatar("https://firefoxusercontent.com/00000000000000000000000000000000", false), "display")
-        )
-
-        profiles.map { it.toDomainObject() }
-            .map { it.avatarSetStrategy }
-            .forEach { assertEquals(defaultProfileAvatarImage, it) }
-    }
-
-    @Test
-    fun `GIVEN profile avatar is nonnull and not default WHEN profile is converted to domain object THEN that url should be used`() {
-        val url = "https://www.mozilla.org"
-        val expectedStrategy = ImageSetStrategy.ByPath(url, placeholderId = R.drawable.ic_default_avatar, errorId = R.drawable.ic_default_avatar)
-        val profiles = listOf(
-            Profile("uid", null, Avatar(url, true), null),
-            Profile("uid", null, Avatar(url, false), null),
-            Profile(null, null, Avatar(url, false), null),
-            Profile("uid", null, Avatar(url, false), "display")
-        )
-
-        profiles.map { it.toDomainObject() }
-            .map { it.avatarSetStrategy }
-            .forEach { assertEquals(expectedStrategy, it) }
     }
 
     @Test
